Renames semi-auto ValueClassCodecs to UnwrappedCodecs
