+++ modules/generic-extras/src/main/scala/io/circe/generic/extras/decoding/UnwrappedDecoder.scala
@@ -3,17 +3,17 @@ package io.circe.generic.extras.decoding
 import io.circe.{Decoder, HCursor}
 import shapeless.{::, Generic, HNil, Lazy}
 
-abstract class ValueClassDecoder[A] extends Decoder[A]
+abstract class UnwrappedDecoder[A] extends Decoder[A]
 
-final object ValueClassDecoder {
-  implicit def decodeValueClass[A <: AnyVal, R](
+final object UnwrappedDecoder {
+  implicit def decodeUnwrapped[A <: AnyVal, R](
     implicit
     gen:    Lazy[Generic.Aux[A, R :: HNil]],
     decode: Decoder[R]
-  ): ValueClassDecoder[A] = new ValueClassDecoder[A] {
+  ): UnwrappedDecoder[A] = new UnwrappedDecoder[A] {
     override def apply(c: HCursor): Decoder.Result[A] =
       decode(c) match {
-        case Right(value) => Right(gen.value.from(value :: HNil))
+        case Right(unwrapped) => Right(gen.value.from(unwrapped :: HNil))
         case l @ Left(_) => l.asInstanceOf[Decoder.Result[A]]
       }
   }
