+++ modules/generic-extras/src/main/scala/io/circe/generic/extras/encoding/UnwrappedEncoder.scala
@@ -3,14 +3,14 @@ package io.circe.generic.extras.encoding
 import io.circe.{Encoder, Json}
 import shapeless.{::, Generic, HNil, Lazy}
 
-abstract class ValueClassEncoder[A] extends Encoder[A]
+abstract class UnwrappedEncoder[A] extends Encoder[A]
 
-final object ValueClassEncoder {
-  implicit def encodeValueClass[A <: AnyVal, R](
+final object UnwrappedEncoder {
+  implicit def encodeUnwrapped[A <: AnyVal, R](
     implicit
     gen:    Lazy[Generic.Aux[A, R :: HNil]],
     encode: Encoder[R]
-  ): ValueClassEncoder[A] = new ValueClassEncoder[A] {
+  ): UnwrappedEncoder[A] = new UnwrappedEncoder[A] {
     override def apply(a: A): Json =
       encode(gen.value.to(a).head)
   }
