+++ modules/generic-extras/src/main/scala/io/circe/generic/extras/semiauto.scala
@@ -1,8 +1,8 @@
 package io.circe.generic.extras
 
 import io.circe.{Decoder, Encoder, ObjectEncoder}
-import io.circe.generic.extras.decoding.{ConfiguredDecoder, EnumerationDecoder, ReprDecoder, ValueClassDecoder}
-import io.circe.generic.extras.encoding.{ConfiguredObjectEncoder, EnumerationEncoder, ValueClassEncoder}
+import io.circe.generic.extras.decoding.{ConfiguredDecoder, EnumerationDecoder, ReprDecoder, UnwrappedDecoder}
+import io.circe.generic.extras.encoding.{ConfiguredObjectEncoder, EnumerationEncoder, UnwrappedEncoder}
 import io.circe.generic.extras.util.RecordToMap
 import io.circe.generic.util.PatchWithOptions
 import shapeless.{Default, HList, LabelledGeneric, Lazy}
@@ -53,12 +53,12 @@ final object semiauto {
   /**
     * Derive a decoder for a value class.
     */
-  def deriveValueClassDecoder[A](implicit decode: Lazy[ValueClassDecoder[A]]): Decoder[A] = decode.value
+  def deriveUnwrappedDecoder[A](implicit decode: Lazy[UnwrappedDecoder[A]]): Decoder[A] = decode.value
 
   /**
     * Derive an encoder for a value class.
     */
-  def deriveValueClassEncoder[A](implicit encode: Lazy[ValueClassEncoder[A]]): Encoder[A] = encode.value
+  def deriveUnwrappedEncoder[A](implicit encode: Lazy[UnwrappedEncoder[A]]): Encoder[A] = encode.value
 
   final class DerivationHelper[A] {
     final def incomplete[P <: HList, C, D <: HList, T <: HList, R <: HList](implicit
